image: reduce size and use DTB from Linux kernel
Reduce size of the image to 256MB, which reduces the amount of time
spent "dd" the files to the SD-card. For a simple Buildroot filesystem
that is plenty of space.

Also use the DTB from Linux kernel instead of the one from imx-mkimage.
With these changes the system boot up fully to a console.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>